+++ packages/notform/src/components/not-field.vue
@@ -38,30 +38,26 @@ const {
  * @returns Promise resolving to the validation result.
  */
 async function validate() {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return await validateField(props.name as any)
+  return await validateField(props.name)
 }
 
-/**
- * Collection of event handlers for common field interactions.
- */
-const methods = {
-  /** Handles focus loss */
+/** @see {@link NotFieldContext.methods} */
+const methods: NotFieldContext['methods'] = {
   onBlur: function () {
     touchField(props.name)
     if (mode === 'eager' || validateOn.includes('blur')) validate()
   },
-  /** Handles value commitment */
+
   onChange: function () {
     dirtyField(props.name)
     if (mode === 'eager' || validateOn.includes('change')) validate()
   },
-  /** Handles streaming input */
+
   onInput: function () {
     dirtyField(props.name)
     if (mode === 'eager' || validateOn.includes('input')) validate()
   },
-  /** Handles field focus */
+
   onFocus: function () {
     dirtyField(props.name)
     if (mode === 'eager' || validateOn.includes('focus')) validate()
@@ -75,17 +71,13 @@ const context = reactive({
   /** Path identifier */
   name: computed(() => props.name),
   /** First error message if any */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  errors: computed(() => getFieldErrors(props.name as any).map(error => error.message)),
+  errors: computed(() => getFieldErrors(props.name).map(error => error.message)),
   /** Interacted status */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  isTouched: computed(() => touchedFields.value.has(props.name as any)),
+  isTouched: computed(() => touchedFields.value.has(props.name)),
   /** Modification status */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  isDirty: computed(() => dirtyFields.value.has(props.name as any)),
+  isDirty: computed(() => dirtyFields.value.has(props.name)),
   /** Current validity status */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  isValid: computed(() => getFieldErrors(props.name as any).length === 0),
+  isValid: computed(() => getFieldErrors(props.name).length === 0),
   /** Manual validation trigger */
   validate,
   /** Interaction methods */