+++ packages/notform/CHANGELOG.md
@@ -1,5 +1,25 @@
 # notform
 
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- [`0702153`](https://github.com/favorodera/notform/commit/07021530b0765ea40bbfda9ad73e9121034b58d3) Thanks [@favorodera](https://github.com/favorodera)! - refactor: simplify `NotMessage` name prop type
+
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- 75bdc81: refactor: cleanup exportted types from `notform`
+  fix: fixed type components type inference issue on `notform-nuxt`
+  - vue@undefined
+
+## 1.0.0-alpha.6
+
+### Patch Changes
+
+- f843ef7: refactor: enhance types and cleanup internal structure
+
 ## 1.0.0-alpha.5
 
 ### Patch Changes